+++ alicloud/resource_alicloud_ots_table.go
@@ -2,10 +2,12 @@ package alicloud
 
 import (
 	"fmt"
+
+	"time"
+
 	"github.com/aliyun/aliyun-tablestore-go-sdk/tablestore"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/helper/schema"
-	"time"
 )
 
 func resourceAlicloudOtsTable() *schema.Resource {
@@ -34,8 +36,8 @@ func resourceAlicloudOtsTable() *schema.Resource {
 							Required: true,
 						},
 						"type": &schema.Schema{
-							Type:         schema.TypeString,
-							Required:     true,
+							Type:     schema.TypeString,
+							Required: true,
 							ValidateFunc: validateAllowedStringValue([]string{
 								string(IntegerType), string(BinaryType), string(StringType)}),
 						},
@@ -59,8 +61,9 @@ func resourceAliyunOtsTableCreate(d *schema.ResourceData, meta interface{}) erro
 	client := meta.(*AliyunClient).otsconn
 
 	tableMeta := new(tablestore.TableMeta)
-	table_name := d.Get("table_name").(string)
-	tableMeta.TableName = table_name
+
+	tableName := d.Get("table_name").(string)
+	tableMeta.TableName = tableName
 
 	for _, primaryKey := range d.Get("primary_key").([]interface{}) {
 		pk := primaryKey.(map[string]interface{})
@@ -80,37 +83,35 @@ func resourceAliyunOtsTableCreate(d *schema.ResourceData, meta interface{}) erro
 
 	_, err := client.CreateTable(createTableRequest)
 	if err != nil {
-		return fmt.Errorf("Failed to create table with error: %s", err)
+		return fmt.Errorf("failed to create table with error: %s", err)
 	}
 
 	// Need to set id before calling read method or terraform.state won't be generated.
-	d.SetId(table_name)
+	d.SetId(tableName)
 	return resourceAliyunOtsTableUpdate(d, meta)
 }
 
 func resourceAliyunOtsTableRead(d *schema.ResourceData, meta interface{}) error {
-	tableName := d.Get("table_name").(string)
-	describ, err := describeOtsTable(tableName, meta)
-
+	tableName := d.Id()
+	describe, err := describeOtsTable(tableName, meta)
 	if err != nil {
 		return fmt.Errorf("failed to describe table with error: %s", err)
 	}
 
-	d.Set("table_name", describ.TableMeta.TableName)
+	d.Set("table_name", describe.TableMeta.TableName)
 
 	var pks []map[string]interface{}
-	keys := describ.TableMeta.SchemaEntry
-	for k, v := range keys {
+	keys := describe.TableMeta.SchemaEntry
+	for _, v := range keys {
 		item := make(map[string]interface{})
-		item["name"] = k
-		item["type"] = v
+		item["name"] = *v.Name
+		item["type"] = convertPrimaryKeyType(*v.Type)
 		pks = append(pks, item)
 	}
 	d.Set("primary_key", pks)
 
-	d.Set("time_to_live", describ.TableOption.TimeToAlive)
-	d.Set("max_version", describ.TableOption.MaxVersion)
-
+	d.Set("time_to_live", describe.TableOption.TimeToAlive)
+	d.Set("max_version", describe.TableOption.MaxVersion)
 	return nil
 }
 
@@ -119,8 +120,8 @@ func resourceAliyunOtsTableUpdate(d *schema.ResourceData, meta interface{}) erro
 	update := false
 
 	updateTableReq := new(tablestore.UpdateTableRequest)
-	table_name := d.Get("table_name").(string)
-	updateTableReq.TableName = table_name
+	tableName := d.Id()
+	updateTableReq.TableName = tableName
 
 	// As the issue of ots sdk, time_to_live and max_version need to be updated together at present.
 	// For the issue, please refer to https://github.com/aliyun/aliyun-tablestore-go-sdk/issues/18
@@ -147,11 +148,11 @@ func resourceAliyunOtsTableUpdate(d *schema.ResourceData, meta interface{}) erro
 }
 
 func resourceAliyunOtsTableDelete(d *schema.ResourceData, meta interface{}) error {
-	tableName := d.Get("table_name").(string)
+	tableName := d.Id()
 	return resource.Retry(2*time.Minute, func() *resource.RetryError {
 		successFlag, err := deleteOtsTable(tableName, meta)
 		if !successFlag {
-			return resource.RetryableError(fmt.Errorf("Delete instance timeout and got an error: %#v.", err))
+			return resource.RetryableError(fmt.Errorf("delete instance timeout and got an error: %#v", err))
 		}
 		return nil
 	})