+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 ## 1.9.2 (Unreleased)
 
+IMPROVEMENTS:
+
+- *New Resource*: _alicloud_ots_table_(([#442](https://github.com/alibaba/terraform-provider/pull/442)))
+- Support load endpoint from environment variable or specified file (([#459](https://github.com/alibaba/terraform-provider/pull/459)))
+- Update example (([#457](https://github.com/alibaba/terraform-provider/pull/457)))
+- Remove terraform/example (([#458](https://github.com/alibaba/terraform-provider/pull/458)))
+
 BUG FIXES:
 
+- Fix deleting forward table not found error (([#457](https://github.com/alibaba/terraform-provider/pull/457)))
 - Fix deleting slb listener error (([#439](https://github.com/alibaba/terraform-provider/pull/439)))
 - Fix creating vswitch error (([#439](https://github.com/alibaba/terraform-provider/pull/439)))